+++ documents/src/commonMain/kotlin/io/github/nomemmurrakh/documents/DocumentSerialization.kt
@@ -1,22 +1,25 @@
 package io.github.nomemmurrakh.documents
 
 import kotlinx.serialization.DeserializationStrategy
+import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.SerializationStrategy
-import kotlinx.serialization.json.Json
+import kotlinx.serialization.cbor.Cbor
 
+@OptIn(ExperimentalSerializationApi::class)
 internal fun <T> encodeDocument(
     documentKey: String,
     value: T,
     serializer: SerializationStrategy<T>,
     storage: Storage,
-    json: Json,
+    cbor: Cbor,
 ) {
-    serializer.serialize(DocumentEncoder(documentKey, storage, json), value)
+    serializer.serialize(DocumentEncoder(documentKey, storage, cbor), value)
 }
 
+@OptIn(ExperimentalSerializationApi::class)
 internal fun <T> decodeDocument(
     documentKey: String,
     deserializer: DeserializationStrategy<T>,
     storage: Storage,
-    json: Json,
-): T = deserializer.deserialize(DocumentDecoder(documentKey, storage, json))
+    cbor: Cbor,
+): T = deserializer.deserialize(DocumentDecoder(documentKey, storage, cbor))